# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Blog API
  version: 1.0.0
  description: |
    API REST para gerenciamento de blog com autenticação JWT.
    
    Esta API permite:
    - Registrar e autenticar usuários
    - Gerenciar usuários (criar, listar, atualizar, deletar)
    - Controle de acesso baseado em roles (USER, ADMIN)
    
    ## Autenticação
    A maioria dos endpoints requer autenticação via token JWT.
    
    Para autenticar:
    1. Faça login usando o endpoint `/api/blog/auth/login`
    2. Use o token retornado no header Authorization: `Bearer {token}`
    
  contact:
    name: Luca5Eckert
    url: https://github.com/Luca5Eckert/blog-api
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Servidor de desenvolvimento
...
paths:
  /api/blog/auth/register:
    post:
      tags:
        - Autenticação
      summary: Registrar novo usuário
      description: Cria um novo usuário no sistema com as credenciais fornecidas
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
            examples:
              exemplo1:
                summary: Criar usuário comum
                value:
                  name: João Silva
                  email: joao@example.com
                  password: SenhaForte123!
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegisterResponse'
              examples:
                exemplo1:
                  summary: Usuário criado
                  value:
                    id: 1
                    name: João Silva
                    email: joao@example.com
                    roleUser: USER
        '400':
          description: Dados inválidos fornecidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/blog/auth/login:
    post:
      tags:
        - Autenticação
      summary: Fazer login
      description: Autentica um usuário e retorna um token JWT para acesso aos endpoints protegidos
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
            examples:
              exemplo1:
                summary: Login de usuário
                value:
                  email: joao@example.com
                  password: SenhaForte123!
      responses:
        '202':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
              examples:
                exemplo1:
                  summary: Login bem-sucedido
                  value:
                    id: 1
                    email: joao@example.com
                    roles:
                      - authority: ROLE_USER
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Insira o token JWT obtido no endpoint /api/blog/auth/login
  schemas:
    UserRegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          maxLength: 30
          description: Nome do usuário
          example: João Silva
        email:
          type: string
          maxLength: 150
          format: email
          description: Email do usuário
          example: joao@example.com
        password:
          type: string
          maxLength: 180
          format: password
          description: Senha do usuário
          example: SenhaForte123!
    UserRegisterResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID do usuário
          example: 1
        name:
          type: string
          maxLength: 30
          description: Nome do usuário
          example: João Silva
        email:
          type: string
          maxLength: 150
          format: email
          description: Email do usuário
          example: joao@example.com
        roleUser:
          $ref: '#/components/schemas/RoleUser'
    UserLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email do usuário
          example: joao@example.com
        password:
          type: string
          format: password
          description: Senha do usuário
          example: SenhaForte123!
    UserLoginResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID do usuário
          example: 1
        email:
          type: string
          format: email
          description: Email do usuário
          example: joao@example.com
        roles:
          type: array
          description: Roles do usuário
          items:
            type: object
            properties:
              authority:
                type: string
                example: ROLE_USER
        token:
          type: string
          description: Token JWT para autenticação
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
    RoleUser:
      type: string
      enum:
        - USER
        - ADMIN
      description: Role do usuário no sistema
      example: USER
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Mensagem de erro
          example: Dados inválidos fornecidos
        status:
          type: integer
          description: Código de status HTTP
          example: 400